+++ quam_builder/architecture/superconducting/qubit/fixed_frequency_transmon.py
@@ -1,7 +1,10 @@
+from typing import Union
 from quam.core import quam_dataclass
-from quam_builder.architecture.superconducting.qubit.base_transmon import BaseTransmon
 
-__all__ = ["FixedFrequencyTransmon"]
+from quam_builder.architecture.superconducting.qubit.base_transmon import BaseTransmon
+from quam_builder.architecture.superconducting.components.zz_drive import ZZDriveIQ, ZZDriveMW
+
+__all__ = ["FixedFrequencyTransmon",  "FixedFrequencyZZDriveTransmon"]
 
 
 @quam_dataclass
@@ -14,3 +17,10 @@ class FixedFrequencyTransmon(BaseTransmon):
     """
 
     pass
+
+
+@quam_dataclass
+class FixedFrequencyZZDriveTransmon(FixedFrequencyTransmon):
+    """Quam Component for flux-tunable features and added Stark ZZ drive."""
+    
+    xy_detuned: Union[ZZDriveMW, ZZDriveIQ] = None