+++ quam_builder/builder/qop_connectivity/create_wiring.py
@@ -35,6 +35,7 @@ def create_wiring(connectivity: Connectivity) -> dict:
                 WiringLineType.RESONATOR,
                 WiringLineType.DRIVE,
                 WiringLineType.FLUX,
+                WiringLineType.DETUNED_DRIVE,
                 WiringLineType.LASER,
                 WiringLineType.SPCM,
             ]: